Convert IrcEventBroker to Typescript
